% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/K1_Cembedding.r
\name{fitk1}
\alias{fitk1}
\title{Fit mixed model for K=1 (one mobility class)}
\usage{
fitk1(timeall, concall, nobs, ttotal, J, sv, size, burnin, thin, tune, nobs_cum,
  sigma2b1, sigma2beta1, a0start, a1start, b1start)
}
\arguments{
\item{timeall}{numeric vector containing the time points at which measurements are available}

\item{concall}{numeric vector containing the measurements at time points specified in argument time}

\item{nobs}{numeric vector indicating how many measurements have been done for each nucleus}

\item{ttotal}{total number of time points at which measurements are available (for all nuclei)}

\item{J}{number of cell nuclei}

\item{sv}{numeric vector containing the parameters of the IG distributions of the random effects}

\item{size}{number of iterations}

\item{burnin}{length of burnin}

\item{thin}{thinning parameter (e.g., if thin=3, only every third iteration is used)}

\item{tune}{at which iterations tuning should be done}

\item{nobs_cum}{numeric vector containing the cumulated number of time points at which measurements are available for all cells, starting with 0}

\item{sigma2b1}{proposal variance for the MH step for b1off (may be different after the fit because of tuning during the fit)}

\item{sigma2beta1}{proposal variance for the MH step for beta1joff (may be different after the fit because of tuning during the fit)}

\item{a0start}{starting value for a0}

\item{a1start}{starting value for a1}

\item{b1start}{starting value for log(b1off)}
}
\value{
timeall, concall, nobs, ttotal, J, sv, size, burnin, thin, tune, nobs_cum, sigma2b1, sigma2beta1, a0start, a1start, b1start: see description above

tau2_alpha0, tau2_alpha1, tau2_beta1: variances of the random effects

sigma2: estimated variance of the error

a0, a1, b1: numeric vectors containing the saved results for the fixed effects a0, a1, log(b1off)

alpha0, alpha1, beta1: numeric vectors containing the saved results for the random effects

acc_b1, acc_beta1: number of accepted proposals in MH step for parameters b1off and beta1joff
}
\description{
Fit mixed model for K=1 (one mobility class)
}

