% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/K2_Cembedding.r
\name{fitk2}
\alias{fitk2}
\title{Fit mixed model for K=2 (two mobility classes)}
\usage{
fitk2(timeall, concall, nobs, ttotal, J, sv, size, burnin, thin, tune, nobs_cum,
  sigma2b1, sigma2beta1, sigma2b2, sigma2beta2, a0start, a1start, b1start,
  a2start, b2start)
}
\arguments{
\item{timeall}{numeric vector containing the time points at which measurements are available}

\item{concall}{numeric vector containing the measurements at time points specified in argument time}

\item{nobs}{numeric vector indicating how many measurements have been done for each nucleus}

\item{ttotal}{total number of time points at which measurements are available (for all nuclei)}

\item{J}{number of cell nuclei}

\item{sv}{numeric vector containing the parameters of the IG distributions of the random effects}

\item{size}{number of iterations}

\item{burnin}{length of burnin}

\item{thin}{thinning parameter (e.g., if thin=3, only every third iteration is used)}

\item{tune}{at which iterations tuning should be done}

\item{nobs_cum}{numeric vector containing the cumulated number of time points at which measurements are available for all cells, starting with 0}

\item{sigma2b1}{proposal variance for the MH step for b1off (may be different after the fit because of tuning during the fit)}

\item{sigma2beta1}{proposal variance for the MH step for beta1joff (may be different after the fit because of tuning during the fit)}

\item{sigma2b2}{proposal variance for the MH step for b2off (may be different after the fit because of tuning during the fit)}

\item{sigma2beta2}{proposal variance for the MH step for beta2joff (may be different after the fit because of tuning during the fit)}

\item{a0start}{starting value for a0}

\item{a1start}{starting value for a1}

\item{b1start}{starting value for log(b1off)}

\item{a2start}{starting value for a2}

\item{b2start}{starting value for log(b2off)}
}
\value{
timeall, concall, nobs, ttotal, J, sv, size, burnin, thin, tune, nobs_cum, sigma2b1, sigma2beta1, sigma2b2, sigma2beta2, a0start, a1start, b1start, a2start, b2start: see description above

tau2_alpha0, tau2_alpha1, tau2_beta1, tau2_alpha2, tau2_beta2: variances of the random effects

sigma2: estimated variance of the error

a0, a1, b1, a2, b2: numeric vectors containing the saved results for the fixed effects a0, a1, log(b1off)

alpha0, alpha1, beta1, alpha2, beta2: numeric vectors containing the saved results for the random effects

acc_b1, acc_beta1, acc_b2, acc_beta2: number of accepted proposals in MH step for parameters b1off, b2off, beta1joff and beta2joff
}
\description{
Fit mixed model for K=2 (two mobility classes)
}

