% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/frapmm-package.r
\docType{package}
\name{frapmm}
\alias{frapmm}
\alias{frapmm-package}
\title{frapmm - Nonlinear mixed model for a series of FRAP images}
\description{
Fit a nonlinear mixed model to data from a series of FRAP images, i.e., FRAP images of various similar nuclei.
}
\details{
Implemented methods are

fitk1: Fit mixed model for K=1 (one mobility class)

fitk2: Fit mixed model for K=2 (two mobility classes)
}
\author{
Martina Feilke \email{martina.feilke@stat.uni-muenchen.de}

Maintainer: Martina Feilke \email{martina.feilke@stat.uni-muenchen.de}
}

